Authentication testest Cart Rest Api

+++ MusicStore/src/main/java/com/musicstore/controller/api/CartRestController.java
@@ -1,6 +1,7 @@
 package com.musicstore.controller.api;
 
 import java.util.List;
+import java.util.Map;
 import java.util.ArrayList; 
 import java.util.Optional;
 
@@ -14,11 +15,22 @@ import org.springframework.web.server.ResponseStatusException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.musicstore.model.AdminBean;
 import com.musicstore.model.CartBean;
+import com.musicstore.model.WebUserBean;
 import com.musicstore.service.DbCartService;
+import com.musicstore.utility.Utility;
+import com.musicstore.service.DbAdminService;
+import com.musicstore.service.DbWebUserService;
 
 @RestController
 public class CartRestController {
+
+	@Autowired
+	private DbAdminService adminService; 
+	
+	@Autowired
+	private DbWebUserService webuserService; 
 	
 	@Autowired
 	private DbCartService cartService; 
@@ -26,28 +38,40 @@ public class CartRestController {
 	public CartRestController() {}
 	
 	@RequestMapping("/musicstore/api/cart")
-	public Iterable<CartBean> getAll(){
+	public Iterable<CartBean> getAll(@RequestBody WebUserBean b){
+		if(!adminService.isAdmin(b))
+		{
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
 		return cartService.getAll();
 	}
 	
 	@RequestMapping("/musicstore/api/cart/{id}")
-	public CartBean getById(@PathVariable int id){
-		Optional<CartBean> cart = cartService.getById(id);
-		if(cart.isEmpty()){
-			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
+	public CartBean getById(@PathVariable int id,@RequestBody WebUserBean b){
+		if(!cartService.getById(id).get().getMail().equals(b.getMail())){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
 		}
-		return cart.get();
+		return cartService.getById(id).get();
 	}
 
 	@RequestMapping(value  ="/musicstore/api/cart", method = RequestMethod.POST)
-	public CartBean create(@RequestBody CartBean p) {
-		return cartService.create(p);
+	public CartBean create(@RequestBody Map<String, Map<String,String>> map) {
+		CartBean cb = Utility.cartDeMap(map.get("topost"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!webuserService.isWebUser(b) || !b.getMail().equals(cb.getMail())){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
+		return cartService.create(cb);
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/cart/{id}", method = RequestMethod.PUT)
-	public CartBean update(@PathVariable int id, @RequestBody CartBean p) {
-		
-		Optional<CartBean> updatedCart= cartService.update(id, p);
+	public CartBean update(@PathVariable int id,@RequestBody Map<String, Map<String,String>> map) {
+		CartBean cb = Utility.cartDeMap(map.get("toput"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!webuserService.isWebUser(b) || !b.getMail().equals(cb.getMail())){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
+		Optional<CartBean> updatedCart= cartService.update(id, cb);
 		if (updatedCart.isEmpty())
 		{
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
@@ -56,11 +80,14 @@ public class CartRestController {
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/cart/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable int id) {
+	public void delete(@PathVariable int id, @RequestBody WebUserBean b) {
+		if(!webuserService.isWebUser(b) || !b.getMail().equals(cartService.getById(id).get().getMail())){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an admin");
+		}
 		Boolean isDeleted = cartService.delete(id);
 		if (isDeleted==false)
 		{
-			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cart not found");
 		}
 	}
